requirements/test.txt
-r base.txt
django_nose
flake8
mock==1.3.0
pytest-cov==2.1.0
pytest-django==2.9.1
pytest-xdist==1.12
pytest==2.8.2
